pyproject.toml
[project]
name = "dharma-agent"
version = "0.1.0"
description = "An A2A agent rooted in the Five Mindfulness Trainings"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]",
    "anthropic",
    "uvicorn",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
